Add service admin for add ticket

+++ posttest/src/main/java/com/kbtg/bootcamp/posttest/admin/AdminAddTicketRequest.java
@@ -11,9 +11,9 @@ public record AdminAddTicketRequest (
                 @Size(min = 6,max = 6)
         String ticket,
 
-        BigInteger price,
+        Integer price,
 
-        BigInteger amount
+        Integer amount
 ) {
 
 }
